% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{algorithm_evaluation}
\alias{algorithm_evaluation}
\title{Title}
\usage{
algorithm_evaluation(code_string, input_data, named_list_results)
}
\arguments{
\item{code_string}{a line of code as a string}

\item{input_data}{an input data frame, or a list of data frames}

\item{named_list_results}{a named list with metrics and values}
}
\value{
an instance of group_comparison class
}
\description{
Title
}
