% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Create a correlation_analysis instance}
\usage{
correlation_analysis(code_string, input_data, test_results)
}
\arguments{
\item{code_string}{A line of code as a string}

\item{input_data}{A data frame, a named list, or a URL as a string}

\item{test_results}{A data frame or a list of data frames}
}
\value{
A correlation_analysis instance
}
\description{
Create a correlation_analysis instance
}
\examples{
res <- data.frame(result_1 = 1, result_2 = 2)
inst_ca <- correlation_analysis(
"stats::cor.test(iris$Petal.Length, iris$Sepal.Length)",
iris,
res
)
}
