% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{data_analysis}
\alias{data_analysis}
\title{Title}
\usage{
data_analysis(instances, code_reference = NULL)
}
\arguments{
\item{instances}{analytic instance or a list of instances}

\item{code_reference}{a URL of the code implementing data analysis}
}
\value{
an instance of data analysis class
}
\description{
Title
}
