% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{descriptive_statistics}
\alias{descriptive_statistics}
\title{Title}
\usage{
descriptive_statistics(code_string, input_data, test_results)
}
\arguments{
\item{code_string}{a line of code as a string}

\item{input_data}{an input data frame, or a list of data frames}

\item{test_results}{a resulting data frame, or a list of data frames}
}
\value{
an instance of regression_analysis class
}
\description{
Title
}
