% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{factor_analysis}
\alias{factor_analysis}
\title{Create a factor_analysis instance}
\usage{
factor_analysis(code_string, input_data, test_results)
}
\arguments{
\item{code_string}{A line of code as a string}

\item{input_data}{A data frame, a named list, or a URL as a string}

\item{test_results}{A data frame or a list of data frames}
}
\value{
A factor_analysis instance
}
\description{
Create a factor_analysis instance
}
\examples{
fa_data <- iris[-5]
res <- data.frame(result_1 = 1, result_2 = 2)
inst_fa <- factor_analysis(
"stats::princomp(fa_data)",
iris,
res
)
}
