% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{group_comparison}
\alias{group_comparison}
\title{Create a group_comparison instance}
\usage{
group_comparison(code_string, input_data, test_results)
}
\arguments{
\item{code_string}{A line of code as a string}

\item{input_data}{A data frame, a named list, or a URL as a string}

\item{test_results}{A data frame or a list of data frames}
}
\value{
A group_comparison instance
}
\description{
Create a group_comparison instance
}
\examples{
res <- data.frame(result_1 = 1, result_2 = 2)
inst_gc <- group_comparison(
"stats::aov(Petal.Length ~ Species, data = iris)",
iris,
res
)
}
