% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leso_12_31_19.R
\docType{data}
\name{leso_12_31_19}
\alias{leso_12_31_19}
\title{1033 Program Data as of 12.31.19}
\format{A data frame with 11 columns, which are:
\describe{
\item{State}{State where department is located}
\item{Station Name (LEA)}{Name of department}
\item{NSN}{Identification number of equipment}
\item{Item Name}{Self explanatory}
\item{Quantity}{Self explanatory}
\item{UI}{Unsure}
\item{Acquisition Value}{Value of equipment, not paid by department}
\item{DEMIL Code}{Military code indicating how equipment must be disposed of}
\item{DEMIL IC}{Unsure}
\item{Ship Date}{Date of transfer}
\item{Station Type}{Unsure}
}}
\usage{
leso_12_31_19
}
\description{
Each piece of transferred equipment is listed, along with the department, location, date of transfer, and acquisition value.
}
\keyword{datasets}
