% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrap-package.R
\docType{package}
\name{mrap-package}
\alias{mrap}
\alias{mrap-package}
\title{mrap: Machine-Readable Data Analysis Results with Function Wrappers}
\description{
You can use the set of wrappers for analytical schemata to reduce the effort in writing machine-readable data. The set of all-in-one wrappers will cover widely used functions from data analysis packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/TIBHannover/lki/knowledge-loom/mrap-r}
  \item Report bugs at \url{https://gitlab.com/TIBHannover/lki/knowledge-loom/mrap-r/-/issues}
}

}
\author{
\strong{Maintainer}: Olga Lezhnina \email{olga.lezhnina@tib.eu} (\href{https://orcid.org/0000-0003-2237-7725}{ORCID})

Authors:
\itemize{
  \item Manuel Prinz \email{manuel.prinz@tib.eu} (\href{https://orcid.org/0000-0003-2151-4556}{ORCID})
  \item Markus Stocker \email{markus.stocker@tib.eu} (\href{https://orcid.org/0000-0001-5492-3212}{ORCID})
}

Other contributors:
\itemize{
  \item Open Research Knowledge Graph Project and Contributors [copyright holder]
}

}
\keyword{internal}
