% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{to_jsonld}
\alias{to_jsonld}
\title{Write an instance in JSON-LD format}
\usage{
to_jsonld(instance)
}
\arguments{
\item{instance}{An instance of an R6 class}
}
\value{
JSON string in JSON-LD format
}
\description{
This function is imported from dtreg for ease-of-use
}
