% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{algorithm_evaluation}
\alias{algorithm_evaluation}
\title{Create an algorithm_evaluation instance}
\usage{
algorithm_evaluation(code_string, input_data, named_list_results)
}
\arguments{
\item{code_string}{A line of code as a string}

\item{input_data}{A data frame, a named list, or a URL as a string}

\item{named_list_results}{A named list with metrics and values}
}
\value{
An algorithm_evaluation instance
}
\description{
Create an algorithm_evaluation instance
}
\examples{
res <- list(F1= 0.46, recall = 0.51)
inst_ae <- algorithm_evaluation("N/A", "data_url", res)
}
