% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{class_discovery}
\alias{class_discovery}
\title{Create a class_discovery instance}
\usage{
class_discovery(code_string, input_data, test_results)
}
\arguments{
\item{code_string}{A line of code as a string, or "N/A" if not given}

\item{input_data}{A data frame, a named list, or a URL as a string}

\item{test_results}{A data frame or a list of data frames}
}
\value{
A class_discovery instance
}
\description{
Create a class_discovery instance
}
\examples{
clust_data <- iris[-5]
res <- data.frame(result_1 = 1, result_2 = 2)
inst_cd <- class_discovery(
"stats::kmeans(clust_data, 3)",
iris,
res
)
}
