% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{data_analysis}
\alias{data_analysis}
\title{Create a data_analysis instance}
\usage{
data_analysis(instances, code_reference = NULL)
}
\arguments{
\item{instances}{Analytic instance or a list of instances}

\item{code_reference}{A URL of the code implementing data analysis}
}
\value{
A data analysis instance
}
\description{
Create a data_analysis instance
}
\examples{
res <- data.frame(mean = 3.758)
inst_ds <- descriptive_statistics(
"base::mean(iris$Petal.Length)",
iris,
res
)
inst_da <- data_analysis(inst_ds)
}
