% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{descriptive_statistics}
\alias{descriptive_statistics}
\title{Create a descriptive_statistics instance}
\usage{
descriptive_statistics(code_string, input_data, test_results)
}
\arguments{
\item{code_string}{A line of code as a string}

\item{input_data}{A data frame, a named list, or a URL as a string}

\item{test_results}{A data frame or a list of data frames}
}
\value{
A descriptive_statistics instance
}
\description{
Create a descriptive_statistics instance
}
