% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_instances.R
\name{multilevel_analysis}
\alias{multilevel_analysis}
\title{Create a multilevel_analysis instance}
\usage{
multilevel_analysis(code_string, input_data, test_results)
}
\arguments{
\item{code_string}{A line of code as a string, or "N/A" if not given}

\item{input_data}{A data frame, a named list, or a URL as a string}

\item{test_results}{A data frame or a list of data frames}
}
\value{
A multilevel_analysis instance
}
\description{
Create a multilevel_analysis instance
}
